+++ tests/Fixtures/Suit.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace RobertStanciu\Wireless\Tests\Fixtures;
+
+enum Suit: string
+{
+    case Hearts = 'hearts';
+    case Spades = 'spades';
+}